moved MQTT code in class; added application setting to disable mqtt
