+++ TeslaLogger/App.config
@@ -31,6 +31,9 @@
             <setting name="SuspendAPIMinutes" serializeAs="String">
                 <value>21</value>
             </setting>
+            <setting name="UseMQTT" serializeAs="String">
+                <value>True</value>
+            </setting>
         </TeslaLogger.ApplicationSettings>
     </applicationSettings>
 </configuration>
