+++ TeslaLogger/Program.cs
@@ -97,22 +97,9 @@
                 UpdateTeslalogger.Start(wh);
                 UpdateTeslalogger.UpdateGrafana(wh);
 
-                try
-                {
-                    System.Threading.Thread MQTTthread = new System.Threading.Thread(StartMqttClient);
-                    MQTTthread.Start();
-                }
-                catch (Exception ex)
-                {
-                    Tools.Log(ex.ToString());
-                }
-
-                // Task.Factory.StartNew(() => wh.UpdateAllPosAddresses());
+                MQTTClient.StartMQTTClient();
+                
                 Task.Factory.StartNew(() => wh.UpdateAllPOIAddresses());
-                // wh.IsDriving();
-                // wh.GetCachedRollupData();
-
-                //wh.GetEnergyChartData();
 
                 while (true)
                 {
@@ -432,45 +419,6 @@
                         Tools.Log("Error getting new Token!");
                     }
                 }
-            }
-        }
-
-        private static void StartMqttClient()
-        {
-            string MQTTClientPath = "/etc/teslalogger/MQTTClient.exe";
-
-            try
-            {
-                if (!System.IO.File.Exists(MQTTClientPath))
-                {
-                    Tools.Log("MQTTClient.exe not found!");
-                    return;
-                }
-
-                System.Diagnostics.Process proc = new System.Diagnostics.Process();
-                proc.EnableRaisingEvents = false;
-                proc.StartInfo.UseShellExecute = false;
-                proc.StartInfo.RedirectStandardOutput = true;
-                proc.StartInfo.FileName = "mono";
-                proc.StartInfo.Arguments = MQTTClientPath;
-
-                proc.Start();
-                
-                while (!proc.StandardOutput.EndOfStream)
-                {
-                    string line = proc.StandardOutput.ReadLine();
-                    Tools.Log(line);
-                }
-
-                proc.WaitForExit();
-            }
-            catch (Exception ex)
-            {
-                Tools.Log(ex.ToString());
-            }
-            finally
-            {
-                Tools.Log("MQTT terminated");
             }
         }
     }
