+++ TeslaLogger/TeslaLogger.csproj
@@ -60,6 +60,7 @@
     <Compile Include="CurrentJSON.cs" />
     <Compile Include="DBViews.cs" />
     <Compile Include="FileManager.cs" />
+    <Compile Include="MQTTClient.cs" />
     <Compile Include="UpdateTeslalogger.cs" />
     <Compile Include="DBHelper.cs" />
     <Compile Include="Geofence.cs" />
